完成API函数
按readme中的格式填充传入的vector
