+++ .gitignore
@@ -4,3 +4,4 @@
 =======
 *.xml
 >>>>>>> 2b0a487c1c606a73b45f8ee079c19fecc065e224
+OpticalDesign
