--- Optical Design .md
@@ -1,8 +1,0 @@
-像差曲线的绘制
-
-用户图形界面
-
-
-
-## 光线追迹程序
-
