+++ README.md
@@ -117,8 +117,8 @@ Optsys(a,nsf,dists[],rs[],ns[]);
 全孔径位置色差 LCAx, 0.7孔径位置色差 LCAxu, 0孔径位置色差 LCAxu0，
 全视场倍率色差 LCAy, 0.7视场倍率色差 LCAyw,
 初级子午场曲 xt', 初级弧矢场曲 xs', 初级像散 xts', 
+全视场绝对畸变 adt, 0.7视场绝对畸变 adtw,
 全视场相对畸变 rdt, 0.7视场相对畸变 rdtw,
-全视场绝对畸变 adt, 0.7视场绝对畸变 adtw
 全市场全孔径、全视场0.7孔径、0.7视场全孔径、0.7视场0.7孔径的子午彗差coma1,coma2,coma3,coma4 
 
 ```
